refactor(externals): modify externals components structure
