+++ packages/mirror-media-next/pages/externals/[partnerSlug].js
@@ -2,7 +2,7 @@ import errors from '@twreporter/errors'
 import styled from 'styled-components'
 
 import client from '../../apollo/apollo-client'
-import ExternalArticles from '../../components/externals/externals-articles'
+import ExternalArticles from '../../components/externals/partner-articles'
 import { GCP_PROJECT_ID } from '../../config/index.mjs'
 import { fetchHeaderDataInDefaultPageLayout } from '../../utils/api'
 import Layout from '../../components/shared/layout'
